Separat fil for Lighthouse tester
